requirements.txt
requests>=2.31
beautifulsoup4>=4.12
pandas>=2.0
openpyxl>=3.1
lxml>=4.9
numpy>=1.24
scipy>=1.11
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Optional — only needed if using --selenium mode
# selenium>=4.15
# webdriver-manager>=4.0
